Use `cabal-doctests` to find all the doctests in `clash-prelude`
