+++ clash-prelude/src/Clash/Explicit/BlockRam.hs
@@ -421,6 +421,8 @@ import           Clash.XException
   (maybeIsX, seqX, NFDataX, deepErrorX, defaultSeqX, fromJustX)
 
 {- $setup
+>>> :m -Clash.Prelude
+>>> :m -Clash.Prelude.Safe
 >>> import Clash.Explicit.Prelude as C
 >>> import qualified Data.List as L
 >>> :set -XDataKinds -XRecordWildCards -XTupleSections -XDeriveAnyClass -XDeriveGeneric
@@ -485,72 +487,72 @@ alu CmpGt x y = if x > y then 1 else 0
 :}
 
 >>> :{
-cpu :: Vec 7 Value          -- ^ Register bank
-    -> (Value,Instruction)  -- ^ (Memory output, Current instruction)
-    -> ( Vec 7 Value
-       , (MemAddr,Maybe (MemAddr,Value),InstrAddr)
-       )
-cpu regbank (memOut,instr) = (regbank',(rdAddr,(,aluOut) <$> wrAddrM,fromIntegral ipntr))
-  where
-    -- Current instruction pointer
-    ipntr = regbank C.!! PC
-    -- Decoder
-    (MachCode {..}) = case instr of
-      Compute op rx ry res -> nullCode {inputX=rx,inputY=ry,result=res,aluCode=op}
-      Branch cr a          -> nullCode {inputX=cr,jmpM=Just a}
-      Jump a               -> nullCode {aluCode=Incr,jmpM=Just a}
-      Load a r             -> nullCode {ldReg=r,rdAddr=a}
-      Store r a            -> nullCode {inputX=r,wrAddrM=Just a}
-      Nop                  -> nullCode
-    -- ALU
-    regX   = regbank C.!! inputX
-    regY   = regbank C.!! inputY
-    aluOut = alu aluCode regX regY
-    -- next instruction
-    nextPC = case jmpM of
-               Just a | aluOut /= 0 -> ipntr + a
-               _                    -> ipntr + 1
-    -- update registers
-    regbank' = replace Zero   0
-             $ replace PC     nextPC
-             $ replace result aluOut
-             $ replace ldReg  memOut
-             $ regbank
-:}
-
->>> :{
-dataMem
-  :: KnownDomain dom
-  => Clock  dom
-  -> Reset  dom
-  -> Enable dom
-  -> Signal dom MemAddr
-  -> Signal dom (Maybe (MemAddr,Value))
-  -> Signal dom Value
-dataMem clk rst en rd wrM = mealy clk rst en dataMemT (C.replicate d32 0) (bundle (rd,wrM))
-  where
-    dataMemT mem (rd,wrM) = (mem',dout)
+let cpu :: Vec 7 Value          -- ^ Register bank
+        -> (Value,Instruction)  -- ^ (Memory output, Current instruction)
+        -> ( Vec 7 Value
+           , (MemAddr,Maybe (MemAddr,Value),InstrAddr)
+           )
+    cpu regbank (memOut,instr) = (regbank',(rdAddr,(,aluOut) <$> wrAddrM,fromIntegral ipntr))
       where
-        dout = mem C.!! rd
-        mem' = case wrM of
-                 Just (wr,din) -> replace wr din mem
-                 Nothing       -> mem
+        -- Current instruction pointer
+        ipntr = regbank C.!! PC
+        -- Decoder
+        (MachCode {..}) = case instr of
+          Compute op rx ry res -> nullCode {inputX=rx,inputY=ry,result=res,aluCode=op}
+          Branch cr a          -> nullCode {inputX=cr,jmpM=Just a}
+          Jump a               -> nullCode {aluCode=Incr,jmpM=Just a}
+          Load a r             -> nullCode {ldReg=r,rdAddr=a}
+          Store r a            -> nullCode {inputX=r,wrAddrM=Just a}
+          Nop                  -> nullCode
+        -- ALU
+        regX   = regbank C.!! inputX
+        regY   = regbank C.!! inputY
+        aluOut = alu aluCode regX regY
+        -- next instruction
+        nextPC = case jmpM of
+                   Just a | aluOut /= 0 -> ipntr + a
+                   _                    -> ipntr + 1
+        -- update registers
+        regbank' = replace Zero   0
+                 $ replace PC     nextPC
+                 $ replace result aluOut
+                 $ replace ldReg  memOut
+                 $ regbank
 :}
 
 >>> :{
-system
-  :: ( KnownDomain dom
-     , KnownNat n )
-  => Vec n Instruction
-  -> Clock dom
-  -> Reset dom
-  -> Enable dom
-  -> Signal dom Value
-system instrs clk rst en = memOut
-  where
-    memOut = dataMem clk rst en rdAddr dout
-    (rdAddr,dout,ipntr) = mealyB clk rst en cpu (C.replicate d7 0) (memOut,instr)
-    instr  = asyncRom instrs <$> ipntr
+let dataMem
+      :: KnownDomain dom
+      => Clock  dom
+      -> Reset  dom
+      -> Enable dom
+      -> Signal dom MemAddr
+      -> Signal dom (Maybe (MemAddr,Value))
+      -> Signal dom Value
+    dataMem clk rst en rd wrM = mealy clk rst en dataMemT (C.replicate d32 0) (bundle (rd,wrM))
+      where
+        dataMemT mem (rd,wrM) = (mem',dout)
+          where
+            dout = mem C.!! rd
+            mem' = case wrM of
+                     Just (wr,din) -> replace wr din mem
+                     Nothing       -> mem
+:}
+
+>>> :{
+let system
+      :: ( KnownDomain dom
+         , KnownNat n )
+      => Vec n Instruction
+      -> Clock dom
+      -> Reset dom
+      -> Enable dom
+      -> Signal dom Value
+    system instrs clk rst en = memOut
+      where
+        memOut = dataMem clk rst en rdAddr dout
+        (rdAddr,dout,ipntr) = mealyB clk rst en cpu (C.replicate d7 0) (memOut,instr)
+        instr  = asyncRom instrs <$> ipntr
 :}
 
 >>> :{
@@ -586,70 +588,70 @@ prog = -- 0 := 4
 :}
 
 >>> :{
-system2
-  :: ( KnownDomain dom
-     , KnownNat n )
-  => Vec n Instruction
-  -> Clock dom
-  -> Reset dom
-  -> Enable dom
-  -> Signal dom Value
-system2 instrs clk rst en = memOut
-  where
-    memOut = asyncRam clk clk en d32 rdAddr dout
-    (rdAddr,dout,ipntr) = mealyB clk rst en cpu (C.replicate d7 0) (memOut,instr)
-    instr  = asyncRom instrs <$> ipntr
+let system2
+      :: ( KnownDomain dom
+         , KnownNat n )
+      => Vec n Instruction
+      -> Clock dom
+      -> Reset dom
+      -> Enable dom
+      -> Signal dom Value
+    system2 instrs clk rst en = memOut
+      where
+        memOut = asyncRam clk clk en d32 rdAddr dout
+        (rdAddr,dout,ipntr) = mealyB clk rst en cpu (C.replicate d7 0) (memOut,instr)
+        instr  = asyncRom instrs <$> ipntr
 :}
 
 >>> :{
-cpu2 :: (Vec 7 Value,Reg)    -- ^ (Register bank, Load reg addr)
-     -> (Value,Instruction)  -- ^ (Memory output, Current instruction)
-     -> ( (Vec 7 Value,Reg)
-        , (MemAddr,Maybe (MemAddr,Value),InstrAddr)
-        )
-cpu2 (regbank,ldRegD) (memOut,instr) = ((regbank',ldRegD'),(rdAddr,(,aluOut) <$> wrAddrM,fromIntegral ipntr))
-  where
-    -- Current instruction pointer
-    ipntr = regbank C.!! PC
-    -- Decoder
-    (MachCode {..}) = case instr of
-      Compute op rx ry res -> nullCode {inputX=rx,inputY=ry,result=res,aluCode=op}
-      Branch cr a          -> nullCode {inputX=cr,jmpM=Just a}
-      Jump a               -> nullCode {aluCode=Incr,jmpM=Just a}
-      Load a r             -> nullCode {ldReg=r,rdAddr=a}
-      Store r a            -> nullCode {inputX=r,wrAddrM=Just a}
-      Nop                  -> nullCode
-    -- ALU
-    regX   = regbank C.!! inputX
-    regY   = regbank C.!! inputY
-    aluOut = alu aluCode regX regY
-    -- next instruction
-    nextPC = case jmpM of
-               Just a | aluOut /= 0 -> ipntr + a
-               _                    -> ipntr + 1
-    -- update registers
-    ldRegD'  = ldReg -- Delay the ldReg by 1 cycle
-    regbank' = replace Zero   0
-             $ replace PC     nextPC
-             $ replace result aluOut
-             $ replace ldRegD memOut
-             $ regbank
+let cpu2 :: (Vec 7 Value,Reg)    -- ^ (Register bank, Load reg addr)
+         -> (Value,Instruction)  -- ^ (Memory output, Current instruction)
+         -> ( (Vec 7 Value,Reg)
+            , (MemAddr,Maybe (MemAddr,Value),InstrAddr)
+            )
+    cpu2 (regbank,ldRegD) (memOut,instr) = ((regbank',ldRegD'),(rdAddr,(,aluOut) <$> wrAddrM,fromIntegral ipntr))
+      where
+        -- Current instruction pointer
+        ipntr = regbank C.!! PC
+        -- Decoder
+        (MachCode {..}) = case instr of
+          Compute op rx ry res -> nullCode {inputX=rx,inputY=ry,result=res,aluCode=op}
+          Branch cr a          -> nullCode {inputX=cr,jmpM=Just a}
+          Jump a               -> nullCode {aluCode=Incr,jmpM=Just a}
+          Load a r             -> nullCode {ldReg=r,rdAddr=a}
+          Store r a            -> nullCode {inputX=r,wrAddrM=Just a}
+          Nop                  -> nullCode
+        -- ALU
+        regX   = regbank C.!! inputX
+        regY   = regbank C.!! inputY
+        aluOut = alu aluCode regX regY
+        -- next instruction
+        nextPC = case jmpM of
+                   Just a | aluOut /= 0 -> ipntr + a
+                   _                    -> ipntr + 1
+        -- update registers
+        ldRegD'  = ldReg -- Delay the ldReg by 1 cycle
+        regbank' = replace Zero   0
+                 $ replace PC     nextPC
+                 $ replace result aluOut
+                 $ replace ldRegD memOut
+                 $ regbank
 :}
 
 >>> :{
-system3
-  :: ( KnownDomain dom
-     , KnownNat n )
-  => Vec n Instruction
-  -> Clock dom
-  -> Reset dom
-  -> Enable dom
-  -> Signal dom Value
-system3 instrs clk rst en = memOut
-  where
-    memOut = blockRam clk en (C.replicate d32 0) rdAddr dout
-    (rdAddr,dout,ipntr) = mealyB clk rst en cpu2 ((C.replicate d7 0),Zero) (memOut,instr)
-    instr  = asyncRom instrs <$> ipntr
+let system3
+      :: ( KnownDomain dom
+         , KnownNat n )
+      => Vec n Instruction
+      -> Clock dom
+      -> Reset dom
+      -> Enable dom
+      -> Signal dom Value
+    system3 instrs clk rst en = memOut
+      where
+        memOut = blockRam clk en (C.replicate d32 0) rdAddr dout
+        (rdAddr,dout,ipntr) = mealyB clk rst en cpu2 ((C.replicate d7 0),Zero) (memOut,instr)
+        instr  = asyncRom instrs <$> ipntr
 :}
 
 >>> :{
