+++ clash-prelude/src/Clash/Explicit/Mealy.hs
@@ -27,6 +27,8 @@ import           Clash.XException      (NFDataX)
 
 {- $setup
 >>> :set -XDataKinds -XTypeApplications
+>>> :m -Clash.Prelude
+>>> :m -Clash.Prelude.Safe
 >>> import Clash.Explicit.Prelude
 >>> import qualified Data.List as L
 >>> :{
