+++ clash-prelude/src/Clash/Explicit/Prelude.hs
@@ -176,6 +176,7 @@ import Clash.XException
 
 {- $setup
 >>> :set -XDataKinds -XTypeApplications
+>>> :m -Clash.Prelude
 >>> import Clash.Explicit.Prelude
 >>> let window4 = window @3
 >>> let windowD3 = windowD @2
