+++ clash-prelude/src/Clash/Prelude/Safe.hs
@@ -156,6 +156,8 @@ import           Clash.XException
 
 {- $setup
 >>> :set -XFlexibleContexts -XTypeApplications
+>>> :m -Clash.Explicit.Prelude
+>>> import Clash.Prelude.Safe
 >>> let rP = registerB (8,8)
 -}
