+++ clash-prelude/src/Clash/Prelude/Testbench.hs
@@ -41,7 +41,11 @@ import Clash.XException                   (ShowX)
 
 {- $setup
 >>> :set -XTemplateHaskell -XDataKinds -XTypeApplications
+>>> :m -Clash.Explicit.Prelude
+>>> :m -Clash.Explicit.Prelude.Safe
+>>> :m -Clash.Explicit.Testbench
 >>> import Clash.Prelude
+>>> import Clash.Prelude.Testbench
 >>> let testInput = stimuliGenerator $(listToVecTH [(1::Int),3..21])
 >>> let expectedOutput = outputVerifier' $(listToVecTH ([70,99,2,3,4,5,7,8,9,10]::[Int]))
 -}
@@ -95,8 +99,8 @@ assertBitVector msg actual expected ret =
 -- testInput = 'stimuliGenerator' $('Clash.Sized.Vector.listToVecTH' [(1::Int),3..21])
 -- @
 --
--- >>> sampleN 13 testInput
--- [1,3,5,7,9,11,13,15,17,19,21,21,21]
+-- >>> sampleN @System 13 testInput
+-- [1,1,3,5,7,9,11,13,15,17,19,21,21]
 stimuliGenerator
   :: ( KnownNat l
      , HiddenClock dom
@@ -120,26 +124,29 @@ stimuliGenerator = hideReset (hideClock E.stimuliGenerator)
 -- @
 --
 -- >>> import qualified Data.List as List
--- >>> sampleN 12 (expectedOutput (fromList ([0..10] List.++ [10,10,10])))
+-- >>> sampleN @System 12 (expectedOutput (fromList (0:[0..10] List.++ [10,10,10])))
 -- <BLANKLINE>
--- cycle(system10000): 0, outputVerifier'
+-- cycle(<Clock: System>): 0, outputVerifier
 -- expected value: 70, not equal to actual value: 0
 -- [False
--- cycle(system10000): 1, outputVerifier'
+-- cycle(<Clock: System>): 1, outputVerifier
+-- expected value: 70, not equal to actual value: 0
+-- ,False
+-- cycle(<Clock: System>): 2, outputVerifier
 -- expected value: 99, not equal to actual value: 1
 -- ,False,False,False,False,False
--- cycle(system10000): 6, outputVerifier'
+-- cycle(<Clock: System>): 7, outputVerifier
 -- expected value: 7, not equal to actual value: 6
 -- ,False
--- cycle(system10000): 7, outputVerifier'
+-- cycle(<Clock: System>): 8, outputVerifier
 -- expected value: 8, not equal to actual value: 7
 -- ,False
--- cycle(system10000): 8, outputVerifier'
+-- cycle(<Clock: System>): 9, outputVerifier
 -- expected value: 9, not equal to actual value: 8
 -- ,False
--- cycle(system10000): 9, outputVerifier'
+-- cycle(<Clock: System>): 10, outputVerifier
 -- expected value: 10, not equal to actual value: 9
--- ,False,True,True]
+-- ,False,True]
 --
 -- If your working with 'BitVector's containing don't care bits you should use 'outputVerifierBitVector''.
 outputVerifier'
