+++ clash-prelude/src/Clash/Signal.hs
@@ -278,22 +278,25 @@ import           Clash.XException      (NFDataX)
 
 {- $setup
 >>> :set -XFlexibleContexts -XTypeApplications
+>>> :m -Clash.Explicit.Prelude
+>>> :m -Clash.Explicit.Prelude.Safe
+>>> import Clash.Prelude
 >>> import Clash.Promoted.Nat (SNat(..))
 >>> import Clash.XException (printX)
 >>> import Control.Applicative (liftA2)
 >>> let oscillate = register False (not <$> oscillate)
 >>> let count = regEn 0 oscillate (count + 1)
 >>> :{
-sometimes1 = s where
-  s = register Nothing (switch <$> s)
-  switch Nothing = Just 1
-  switch _       = Nothing
+let sometimes1 = s where
+      s = register Nothing (switch <$> s)
+      switch Nothing = Just 1
+      switch _       = Nothing
 :}
 
 >>> :{
-countSometimes = s where
-  s     = regMaybe 0 (plusM (pure <$> s) sometimes1)
-  plusM = liftA2 (liftA2 (+))
+let countSometimes = s where
+      s     = regMaybe 0 (plusM (pure <$> s) sometimes1)
+      plusM = liftA2 (liftA2 (+))
 :}
 
 -}
@@ -521,6 +524,7 @@ exposeClock = \f clk -> exposeSpecificClock (const f) clk (Proxy @dom)
 --
 -- <Clash-Signal.html#hiddenclockandreset Click here to read more about hidden clocks, resets, and enables>
 --
+#ifdef CLASH_MULTIPLE_HIDDEN
 -- === __Example__
 -- 'exposeSpecificClock' can only be used when it can find the specified domain
 -- in /r/:
@@ -536,6 +540,7 @@ exposeClock = \f clk -> exposeSpecificClock (const f) clk (Proxy @dom)
 -- reg = 'register' @@dom 5 (reg + 1)
 -- sig = exposeSpecificClock @@dom reg 'clockGen'
 -- @
+#endif
 --
 exposeSpecificClock
    :: forall dom  r
@@ -716,6 +721,7 @@ exposeReset = \f rst -> exposeSpecificReset (const f) rst (Proxy @dom)
 --
 -- <Clash-Signal.html#hiddenclockandreset Click here to read more about hidden clocks, resets, and enables>
 --
+#ifdef CLASH_MULTIPLE_HIDDEN
 -- === __Example__
 -- 'exposeSpecificReset' can only be used when it can find the specified domain
 -- in /r/:
@@ -731,6 +737,7 @@ exposeReset = \f rst -> exposeSpecificReset (const f) rst (Proxy @dom)
 -- reg = 'register' @@dom 5 (reg + 1)
 -- sig = exposeSpecificReset @@dom reg 'resetGen'
 -- @
+#endif
 --
 exposeSpecificReset
   :: forall dom r
@@ -910,6 +917,7 @@ exposeEnable = \f gen -> exposeSpecificEnable (const f) gen (Proxy @dom)
 --
 -- <Clash-Signal.html#hiddenclockandreset Click here to read more about hidden clocks, resets, and enables>
 --
+#ifdef CLASH_MULTIPLE_HIDDEN
 -- === __Example__
 -- 'exposeSpecificEnable' can only be used when it can find the specified domain
 -- in /r/:
@@ -925,6 +933,7 @@ exposeEnable = \f gen -> exposeSpecificEnable (const f) gen (Proxy @dom)
 -- reg = 'register' @@dom 5 (reg + 1)
 -- sig = exposeSpecificEnable @@dom reg 'enableGen'
 -- @
+#endif
 --
 exposeSpecificEnable
   :: forall dom r
