+++ clash-prelude/src/Clash/Signal/Delayed/Internal.hs
@@ -48,6 +48,9 @@ import Clash.XException           (NFDataX)
 {- $setup
 >>> :set -XDataKinds
 >>> :set -XTypeOperators
+>>> :m -Clash.Prelude
+>>> :m -Clash.Prelude.Safe
+>>> :m -Clash.Signal
 >>> import Clash.Explicit.Prelude
 >>> :{
 let mac :: Clock System
