+++ clash-prelude/src/Clash/Signal/Internal.hs
@@ -161,6 +161,9 @@ import Clash.XException
 >>> :set -XDataKinds
 >>> :set -XMagicHash
 >>> :set -XTypeApplications
+>>> :m -Clash.Prelude
+>>> :m -Clash.Signal
+>>> import Clash.Signal.Internal
 >>> import Clash.Promoted.Nat
 >>> import Clash.XException
 >>> type System = "System"
